+++ src/App.tsx
@@ -1,60 +1,22 @@
-import {
-  BellIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-} from '@chakra-ui/icons';
-import {
-  Alert,
-  AlertDialog,
-  AlertDialogBody,
-  AlertDialogContent,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogOverlay,
-  AlertIcon,
-  AlertTitle,
-  Box,
-  Button,
-  CloseButton,
-  Flex,
-  Heading,
-  HStack,
-  IconButton,
-  Select,
-  Table,
-  Tbody,
-  Td,
-  Text,
-  Th,
-  Thead,
-  Tr,
-  VStack,
-} from '@chakra-ui/react';
+import { Box, Flex, } from '@chakra-ui/react';
 import { useRef, useState } from 'react';
+
 import { EventForm } from './components/EventForm.tsx';
 import { EventList } from './components/EventList.tsx';
+import { CalendarView } from './components/CalendarView.tsx';
+import { EventDialog } from './components/EventDialog.tsx';
+import { NotificationList } from './components/NotificationList.tsx';
 import { useCalendarView } from './hooks/useCalendarView.ts';
 import { useEventOperations } from './hooks/useEventOperations.ts';
 import { useEventManagement } from './hooks/useEventManagement.ts';
 import { useNotifications } from './hooks/useNotifications.ts';
 import { useSearch } from './hooks/useSearch.ts';
 import { Event, EventForm as EventFormType } from './types';
-import { weekDays } from './constants/options.ts';
-import {
-  formatDate,
-  formatMonth,
-  formatWeek,
-  getEventsForDay,
-  getWeekDates,
-  getWeeksAtMonth,
-} from './utils/dateUtils';
-
 
 function App() {
   const [isOverlapDialogOpen, setIsOverlapDialogOpen] = useState(false);
   const [overlappingEvents, setOverlappingEvents] = useState<Event[]>([]);
   const [eventToSave, setEventToSave] = useState<Event | EventFormType | null>(null);
-  // const [editingEvent, setEditingEvent] = useState<Event | null>(null);
   const cancelRef = useRef<HTMLButtonElement>(null);
 
   const { editingEvent, editEvent, setEditingEvent } = useEventManagement({});
@@ -87,130 +49,134 @@ function App() {
     deleteEvent(id);
   };
 
-  const renderWeekView = () => {
-    const weekDates = getWeekDates(currentDate);
-    return (
-      <VStack data-testid="week-view" align="stretch" w="full" spacing={4}>
-        <Heading size="md">{formatWeek(currentDate)}</Heading>
-        <Table variant="simple" w="full">
-          <Thead>
-            <Tr>
-              {weekDays.map((day) => (
-                <Th key={day} width="14.28%">
-                  {day}
-                </Th>
-              ))}
-            </Tr>
-          </Thead>
-          <Tbody>
-            <Tr>
-              {weekDates.map((date) => (
-                <Td key={date.toISOString()} height="100px" verticalAlign="top" width="14.28%">
-                  <Text fontWeight="bold">{date.getDate()}</Text>
-                  {filteredEvents
-                    .filter((event) => new Date(event.date).toDateString() === date.toDateString())
-                    .map((event) => {
-                      const isNotified = notifiedEvents.includes(event.id);
-                      return (
-                        <Box
-                          key={event.id}
-                          p={1}
-                          my={1}
-                          bg={isNotified ? 'red.100' : 'gray.100'}
-                          borderRadius="md"
-                          fontWeight={isNotified ? 'bold' : 'normal'}
-                          color={isNotified ? 'red.500' : 'inherit'}
-                        >
-                          <HStack spacing={1}>
-                            {isNotified && <BellIcon />}
-                            <Text fontSize="sm" noOfLines={1}>
-                              {event.title}
-                            </Text>
-                          </HStack>
-                        </Box>
-                      );
-                    })}
-                </Td>
-              ))}
-            </Tr>
-          </Tbody>
-        </Table>
-      </VStack>
-    );
+  const handleRemoveNotification = (index: number) => {
+    setNotifications((prev) => prev.filter((_, i) => i !== index));
   };
 
-  const renderMonthView = () => {
-    const weeks = getWeeksAtMonth(currentDate);
+  // const renderWeekView = () => {
+  //   const weekDates = getWeekDates(currentDate);
+  //   return (
+  //     <VStack data-testid="week-view" align="stretch" w="full" spacing={4}>
+  //       <Heading size="md">{formatWeek(currentDate)}</Heading>
+  //       <Table variant="simple" w="full">
+  //         <Thead>
+  //           <Tr>
+  //             {weekDays.map((day) => (
+  //               <Th key={day} width="14.28%">
+  //                 {day}
+  //               </Th>
+  //             ))}
+  //           </Tr>
+  //         </Thead>
+  //         <Tbody>
+  //           <Tr>
+  //             {weekDates.map((date) => (
+  //               <Td key={date.toISOString()} height="100px" verticalAlign="top" width="14.28%">
+  //                 <Text fontWeight="bold">{date.getDate()}</Text>
+  //                 {filteredEvents
+  //                   .filter((event) => new Date(event.date).toDateString() === date.toDateString())
+  //                   .map((event) => {
+  //                     const isNotified = notifiedEvents.includes(event.id);
+  //                     return (
+  //                       <Box
+  //                         key={event.id}
+  //                         p={1}
+  //                         my={1}
+  //                         bg={isNotified ? 'red.100' : 'gray.100'}
+  //                         borderRadius="md"
+  //                         fontWeight={isNotified ? 'bold' : 'normal'}
+  //                         color={isNotified ? 'red.500' : 'inherit'}
+  //                       >
+  //                         <HStack spacing={1}>
+  //                           {isNotified && <BellIcon />}
+  //                           <Text fontSize="sm" noOfLines={1}>
+  //                             {event.title}
+  //                           </Text>
+  //                         </HStack>
+  //                       </Box>
+  //                     );
+  //                   })}
+  //               </Td>
+  //             ))}
+  //           </Tr>
+  //         </Tbody>
+  //       </Table>
+  //     </VStack>
+  //   );
+  // };
 
-    return (
-      <VStack data-testid="month-view" align="stretch" w="full" spacing={4}>
-        <Heading size="md">{formatMonth(currentDate)}</Heading>
-        <Table variant="simple" w="full">
-          <Thead>
-            <Tr>
-              {weekDays.map((day) => (
-                <Th key={day} width="14.28%">
-                  {day}
-                </Th>
-              ))}
-            </Tr>
-          </Thead>
-          <Tbody>
-            {weeks.map((week, weekIndex) => (
-              <Tr key={weekIndex}>
-                {week.map((day, dayIndex) => {
-                  const dateString = day ? formatDate(currentDate, day) : '';
-                  const holiday = holidays[dateString];
+  // const renderMonthView = () => {
+  //   const weeks = getWeeksAtMonth(currentDate);
 
-                  return (
-                    <Td
-                      key={dayIndex}
-                      height="100px"
-                      verticalAlign="top"
-                      width="14.28%"
-                      position="relative"
-                    >
-                      {day && (
-                        <>
-                          <Text fontWeight="bold">{day}</Text>
-                          {holiday && (
-                            <Text color="red.500" fontSize="sm">
-                              {holiday}
-                            </Text>
-                          )}
-                          {getEventsForDay(filteredEvents, day).map((event) => {
-                            const isNotified = notifiedEvents.includes(event.id);
-                            return (
-                              <Box
-                                key={event.id}
-                                p={1}
-                                my={1}
-                                bg={isNotified ? 'red.100' : 'gray.100'}
-                                borderRadius="md"
-                                fontWeight={isNotified ? 'bold' : 'normal'}
-                                color={isNotified ? 'red.500' : 'inherit'}
-                              >
-                                <HStack spacing={1}>
-                                  {isNotified && <BellIcon />}
-                                  <Text fontSize="sm" noOfLines={1}>
-                                    {event.title}
-                                  </Text>
-                                </HStack>
-                              </Box>
-                            );
-                          })}
-                        </>
-                      )}
-                    </Td>
-                  );
-                })}
-              </Tr>
-            ))}
-          </Tbody>
-        </Table>
-      </VStack>
-    );
-  };
+  //   return (
+  //     <VStack data-testid="month-view" align="stretch" w="full" spacing={4}>
+  //       <Heading size="md">{formatMonth(currentDate)}</Heading>
+  //       <Table variant="simple" w="full">
+  //         <Thead>
+  //           <Tr>
+  //             {weekDays.map((day) => (
+  //               <Th key={day} width="14.28%">
+  //                 {day}
+  //               </Th>
+  //             ))}
+  //           </Tr>
+  //         </Thead>
+  //         <Tbody>
+  //           {weeks.map((week, weekIndex) => (
+  //             <Tr key={weekIndex}>
+  //               {week.map((day, dayIndex) => {
+  //                 const dateString = day ? formatDate(currentDate, day) : '';
+  //                 const holiday = holidays[dateString];
+
+  //                 return (
+  //                   <Td
+  //                     key={dayIndex}
+  //                     height="100px"
+  //                     verticalAlign="top"
+  //                     width="14.28%"
+  //                     position="relative"
+  //                   >
+  //                     {day && (
+  //                       <>
+  //                         <Text fontWeight="bold">{day}</Text>
+  //                         {holiday && (
+  //                           <Text color="red.500" fontSize="sm">
+  //                             {holiday}
+  //                           </Text>
+  //                         )}
+  //                         {getEventsForDay(filteredEvents, day).map((event) => {
+  //                           const isNotified = notifiedEvents.includes(event.id);
+  //                           return (
+  //                             <Box
+  //                               key={event.id}
+  //                               p={1}
+  //                               my={1}
+  //                               bg={isNotified ? 'red.100' : 'gray.100'}
+  //                               borderRadius="md"
+  //                               fontWeight={isNotified ? 'bold' : 'normal'}
+  //                               color={isNotified ? 'red.500' : 'inherit'}
+  //                             >
+  //                               <HStack spacing={1}>
+  //                                 {isNotified && <BellIcon />}
+  //                                 <Text fontSize="sm" noOfLines={1}>
+  //                                   {event.title}
+  //                                 </Text>
+  //                               </HStack>
+  //                             </Box>
+  //                           );
+  //                         })}
+  //                       </>
+  //                     )}
+  //                   </Td>
+  //                 );
+  //               })}
+  //             </Tr>
+  //           ))}
+  //         </Tbody>
+  //       </Table>
+  //     </VStack>
+  //   );
+  // };
 
   return (
     <Box w="full" h="100vh" m="auto" p={5}>
@@ -221,33 +187,16 @@ function App() {
           saveEvent={saveEvent}
           initialEvent={editingEvent}
         />
-        <VStack flex={1} spacing={5} align="stretch">
-          <Heading>일정 보기</Heading>
-
-          <HStack mx="auto" justifyContent="space-between">
-            <IconButton
-              aria-label="Previous"
-              icon={<ChevronLeftIcon />}
-              onClick={() => navigate('prev')}
-            />
-            <Select
-              aria-label="view"
-              value={view}
-              onChange={(e) => setView(e.target.value as 'week' | 'month')}
-            >
-              <option value="week">Week</option>
-              <option value="month">Month</option>
-            </Select>
-            <IconButton
-              aria-label="Next"
-              icon={<ChevronRightIcon />}
-              onClick={() => navigate('next')}
-            />
-          </HStack>
-
-          {view === 'week' && renderWeekView()}
-          {view === 'month' && renderMonthView()}
-        </VStack>
+        
+        <CalendarView 
+          view={view}
+          currentDate={currentDate}
+          filteredEvents={filteredEvents}
+          notifiedEvents={notifiedEvents}
+          holidays={holidays}
+          onViewChange={setView}
+          onNavigate={navigate}
+        />
 
         <EventList 
           events={filteredEvents} 
@@ -259,58 +208,18 @@ function App() {
         />
       </Flex>
 
-      <AlertDialog
+      <EventDialog
         isOpen={isOverlapDialogOpen}
-        leastDestructiveRef={cancelRef}
         onClose={() => setIsOverlapDialogOpen(false)}
-      >
-        <AlertDialogOverlay>
-          <AlertDialogContent>
-            <AlertDialogHeader fontSize="lg" fontWeight="bold">
-              일정 겹침 경고
-            </AlertDialogHeader>
+        onConfirm={handleDialogConfirm}
+        overlappingEvents={overlappingEvents}
+        cancelRef={cancelRef}
+      />
 
-            <AlertDialogBody>
-              다음 일정과 겹칩니다:
-              {overlappingEvents.map((event) => (
-                <Text key={event.id}>
-                  {event.title} ({event.date} {event.startTime}-{event.endTime})
-                </Text>
-              ))}
-              계속 진행하시겠습니까?
-            </AlertDialogBody>
-
-            <AlertDialogFooter>
-              <Button ref={cancelRef} onClick={() => setIsOverlapDialogOpen(false)}>
-                취소
-              </Button>
-              <Button
-                colorScheme="red"
-                onClick={handleDialogConfirm}
-                ml={3}
-              >
-                계속 진행
-              </Button>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialogOverlay>
-      </AlertDialog>
-
-      {notifications.length > 0 && (
-        <VStack position="fixed" top={4} right={4} spacing={2} align="flex-end">
-          {notifications.map((notification, index) => (
-            <Alert key={index} status="info" variant="solid" width="auto">
-              <AlertIcon />
-              <Box flex="1">
-                <AlertTitle fontSize="sm">{notification.message}</AlertTitle>
-              </Box>
-              <CloseButton
-                onClick={() => setNotifications((prev) => prev.filter((_, i) => i !== index))}
-              />
-            </Alert>
-          ))}
-        </VStack>
-      )}
+      <NotificationList 
+        notifications={notifications} 
+        onRemove={handleRemoveNotification}
+      />
     </Box>
   );
 }
